structure

+++ src/main/java/tn/esprit/spring/DAO/Entities/MonthlyPayment.java
@@ -10,7 +10,7 @@ public class MonthlyPayment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long numPayment;
+    int numPayment;
 
     Date paymentDate;
     @ManyToOne
